src/BD3-Dataset/requirements.txt
torch
torchvision
torchaudio

timm
opencv-python
scikit-learn
matplotlib
pandas
numpy
tqdm
Pillow
